+++ app/features/frs/JoinFrsController.scala
@@ -14,39 +14,62 @@
  * limitations under the License.
  */
 
-package controllers.frs
+package models.view.frs {
 
-import javax.inject.Inject
+  import models._
+  import models.api.VatScheme
+  import play.api.libs.json.Json
 
-import cats.syntax.FlatMapSyntax
-import controllers.{CommonPlayDependencies, VatRegistrationController}
-import forms.genericForms.{YesOrNoAnswer, YesOrNoFormFactory}
-import models._
-import models.view.frs.JoinFrsView
-import play.api.mvc.{Action, AnyContent}
-import services.{S4LService, VatRegistrationService}
+  final case class JoinFrsView(selection: Boolean)
 
+  object JoinFrsView {
+    implicit val format = Json.format[JoinFrsView]
 
-class JoinFrsController @Inject()(ds: CommonPlayDependencies, formFactory: YesOrNoFormFactory)
-                                 (implicit s4LService: S4LService, vrs: VatRegistrationService)
-  extends VatRegistrationController(ds) with FlatMapSyntax {
+    implicit val viewModelFormat = ViewModelFormat(
+      readF = (_: S4LFlatRateScheme).joinFrs,
+      updateF = (c: JoinFrsView, g: Option[S4LFlatRateScheme]) =>
+        g.getOrElse(S4LFlatRateScheme()).copy(joinFrs = Some(c))
+    )
 
-  val form = formFactory.form("joinFrs")("frs.join")
+    implicit val modelTransformer = ApiModelTransformer[JoinFrsView] { (vs: VatScheme) =>
+      vs.vatFlatRateScheme.map(_.joinFrs).map(JoinFrsView(_))
+    }
+  }
+}
 
-  def show: Action[AnyContent] = authorised.async(implicit user => implicit request =>
-    viewModel[JoinFrsView]().map(vm => YesOrNoAnswer(vm.selection)).fold(form)(form.fill)
-      .map(f => Ok(features.frs.views.html.frs_join(f))))
+package controllers.frs {
 
-  def submit: Action[AnyContent] = authorised.async(implicit user => implicit request =>
-    form.bindFromRequest().fold(
-      badForm => BadRequest(features.frs.views.html.frs_join(badForm)).pure,
-      view => (if (view.answer) {
-        save(JoinFrsView(view.answer)).map(_ =>
-          controllers.frs.routes.AnnualCostsInclusiveController.show())
-      } else {
-        for {
-          _ <- s4LService.save(S4LFlatRateScheme(joinFrs = Some(JoinFrsView(false))))
-          _ <- vrs.submitVatFlatRateScheme()
-        } yield controllers.routes.SummaryController.show()
-      }).map(Redirect)))
+  import javax.inject.Inject
+
+  import cats.syntax.FlatMapSyntax
+  import controllers.{CommonPlayDependencies, VatRegistrationController}
+  import forms.genericForms.{YesOrNoAnswer, YesOrNoFormFactory}
+  import models._
+  import models.view.frs.JoinFrsView
+  import play.api.mvc.{Action, AnyContent}
+  import services.{S4LService, VatRegistrationService}
+
+  class JoinFrsController @Inject()(ds: CommonPlayDependencies, formFactory: YesOrNoFormFactory)
+                                   (implicit s4LService: S4LService, vrs: VatRegistrationService)
+    extends VatRegistrationController(ds) with FlatMapSyntax {
+
+    val form = formFactory.form("joinFrs")("frs.join")
+
+    def show: Action[AnyContent] = authorised.async(implicit user => implicit request =>
+      viewModel[JoinFrsView]().map(vm => YesOrNoAnswer(vm.selection)).fold(form)(form.fill)
+        .map(f => Ok(features.frs.views.html.frs_join(f))))
+
+    def submit: Action[AnyContent] = authorised.async(implicit user => implicit request =>
+      form.bindFromRequest().fold(
+        badForm => BadRequest(features.frs.views.html.frs_join(badForm)).pure,
+        view => (if (view.answer) {
+          save(JoinFrsView(view.answer)).map(_ =>
+            controllers.frs.routes.AnnualCostsInclusiveController.show())
+        } else {
+          for {
+            _ <- s4LService.save(S4LFlatRateScheme(joinFrs = Some(JoinFrsView(false))))
+            _ <- vrs.submitVatFlatRateScheme()
+          } yield controllers.routes.SummaryController.show()
+        }).map(Redirect)))
+  }
 }
